requirements.txt
nats-py==2.9.0
Pillow==11.1.0
boto3==1.36.0
psycopg[binary]==3.2.4
redis==5.2.1
structlog==24.4.0
